py/data/airflow-notes/pyproject.toml
[project]
name = 'airflow-notes'
version = "0.1.0"
# Airflow doesn't have Python 3.13 support yet
requires-python = "==3.12.*"
dependencies = [
    "apache-airflow>=2.10.2",
    "pandas>=2.2.3",
    "typing-extensions>=4.12.2",
]

tool.uv.dev-dependencies = ["ruff", "pylint"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.pyright]
typeCheckingMode = "basic"
